refactor operator UI shared helpers
